#44 [REFACOTR] 필터 함수 UTILS 폴더로 이동
